A model with class 😎
Created the TFSocketModel.py file with the TFSocketModel class.

Code runs but the model is no longer learning.
