+++ src/agents/tfsocket/TFSocket.py
@@ -3,8 +3,7 @@ import socket
 import os
 import random
 import traceback
-import contextlib
-import tensorflow as tf
+import TFSocketModel as tfmodel
 
 #This must be kept in sync with the Java side!
 # Define global constants
@@ -34,157 +33,8 @@ def send_letter(conn, letter):
     conn.sendall(letter.encode('ASCII'))
     return letter
 
-def calc_avg_steps(history):
-    '''
-    Calculate average steps to goal for a given history
-    '''
-    # Count the number of goals (skipping the first)
-    num_historical_goals = 0
-    for i in range(len(history)):
-        if (i == 0):
-            continue
-        if (history[i].isupper()):
-            num_historical_goals+=1
-    
-    avg_steps = len(history) / num_historical_goals
-    log(f'Avg steps: {avg_steps}')
-    return avg_steps
-
-   
-def create_model(entire_history):
-    '''
-    Train and optimize a new ANN model
-    
-    Refer to the beginner tensorflow tutorial:
-    https://www.tensorflow.org/tutorials/quickstart/beginner
-    
-    Global variables being used:
-        model
-    '''
-    global model
-
-    # Defining the model function as a variable
-    model = tf.keras.models.Sequential([
-        tf.keras.layers.Dense(16, activation='relu'),
-        tf.keras.layers.Dropout(0.2),
-        tf.keras.layers.Dense(2)
-    ])
-
-    # Defining the inputs and expected outputs from a given history
-    avg_steps = calc_avg_steps(entire_history)
-    x_train = calc_input_tensors(entire_history)
-    y_train = calc_desired_actions(entire_history, avg_steps)
-    x_train = tf.constant(x_train)
-    y_train = tf.constant(y_train)
-    
-    # Defining the model's loss function
-    predictions = model(x_train[:1])
-    predictions = tf.nn.softmax(predictions).numpy() # Convert to probabilities
-    loss_fn = tf.keras.losses.SparseCategoricalCrossentropy(from_logits=True)
-
-    # Optimize and train the model
-    model.compile(optimizer='adam', loss=loss_fn, metrics=['accuracy'])
-    with contextlib.redirect_stdout(open('trainout.txt', 'w')):
-        model.fit(x_train, y_train, epochs=5, verbose=2)
-
-    return model, avg_steps
-
-def calc_input_tensors(history):
-    '''
-    Convert a blind FSM action history into an array of input tensors
-    '''
-    # Sanity check
-    if (len(history) < WINDOW_SIZE):
-        log('ERROR:  history too short for training!')
-        return
-
-    # Iterate over a range that skips the first WINDOW_SIZE indexes so we have a full window for each input
-    # generate an input tensor for each one
-    history_range = list(range(len(history)))[WINDOW_SIZE:]
-    tensor = []
-    for i in history_range:
-        window = history[i-WINDOW_SIZE:i]
-        tensor.append(flatten(window))
-
-    return tensor
-
-##
-# calc_desired_actions
-#
-# the output of this function should line up with the output of 
-# calc_input_tensors for the same history
-#
-# history - in a letter-based format (e.g., aabaBbbabaabaBbbab...)
-# cutoff - distance from curr pos to goal at which the agent prefers 
-#          NOT to take the historical action
-def calc_desired_actions(history, cutoff):
-    '''
-    Calculates the desired action for each window in a history
-    '''
-    #sanity check
-    if (len(history) < WINDOW_SIZE):
-        log('ERROR:  history too short for training!')
-        return
-    
-    #iterate over a range starting at index WINDOW_SIZE in the history
-    hrange = list(range(len(history)))[WINDOW_SIZE:]
-    ret = []
-    for i in hrange:
-        #calculate how many steps from this position to the next goal
-        num_steps = 0
-        while(history[i + num_steps].islower()):
-            num_steps += 1
-        
-        #calculate the index of the action of this position
-        val = 0
-        if ( (history[i].lower() == 'b') or (history[i].lower() == 'B') ):
-            val = 1
-            
-        #if the cutoff is exceeded then we want the other action instead
-        if (num_steps >= cutoff):
-            val = 1 - val
-            
-        #add to result
-        ret.append(val)
-    
-    return ret
-            
-def flatten(window):
-    '''
-    Convert a window into an input tensor
-    Helper function for @calc_input_tensors()
-    
-    A window has this format:      'aabaBbbab'
-    the return object's length is 3*WINDOW_SIZE
-    being whether or not an 'a'/'A' is in that position
-    the second set is for b/B.  The third set is goal sensor.
-
-    TODO:  This is hard-coded for a two-letter ALPHABET. Change
-           the code to handle any ALPHABET size
-    TODO:  In the future we may want to handle sensors. 
-           Example input window:   '01a11B00b00a'
-    '''
-    win_size = len(window)
-
-    # Split the window into goal part and action part
-    ays = [0.0 for i in range(win_size)]
-    bees = [0.0 for i in range(win_size)]
-    goals = [0.0 for i in range(win_size)]
-
-    for i in range(win_size):
-        let = window[i]
-        if let == 'a' or let == 'A':
-            ays[i] = 1.0
-        else:
-            bees[i] = 1.0
-        if let == 'A' or let == 'B':
-            goals[i] = 1.0
-
-    return ays + bees + goals
-
-def check_if_goal(history, steps_since_last_goal, entire_history):
+def check_if_goal(history, steps_since_last_goal, entire_history, model):
     ''' Did we reach the goal? '''
-    global model
     num_goals = 0
     avg_steps = 0
 
@@ -194,39 +44,18 @@ def check_if_goal(history, steps_since_last_goal, entire_history):
         log(f'Found goal #{num_goals}')
         if ((num_goals >= 400) and (model is None)):
             log('Creating model, reached 400 goals')
-            _, avg_steps = create_model(entire_history)
+            model = tfmodel(WINDOW_SIZE, entire_history)
+            avg_steps = model.train_model(entire_history)
     else:
         steps_since_last_goal +=1
 
-    return steps_since_last_goal, avg_steps
+    return model, steps_since_last_goal, avg_steps
 
-def get_model_prediction(history):
-    '''
-    Uses the model to generate a next step for the environment
-    '''
-    global model
-
-    one_input = []
-    one_input.append(flatten(history))
-    one_input = tf.constant(one_input)
-    predictions = model(one_input)
-
-    last_step = 'a'
-    # Predictions now returns as a tensor of shape (1,2)
-    log(f'Chance of a: {predictions[0][0]}\nChance of b: {predictions[0][1]}')
-    if (predictions[0][0] < predictions[0][1]):
-        last_step = 'b'
-    log('Model is not None, sending prediction: ' + last_step)
-    
-    return last_step
-
-def process_history_sentinel(strData, entire_history, steps_since_last_goal):
+def process_history_sentinel(strData, entire_history, steps_since_last_goal, model):
     '''
     Manage the step-history and use the model to generate the next
     step for the environment
     '''
-    global model
-    
     # Update entire_history
     history = strData[11:]
     log(f'Window received: {history}')
@@ -234,7 +63,7 @@ def process_history_sentinel(strData, entire_history, steps_since_last_goal):
         history = history.lower()
     updated_entire_history = entire_history + str(history[-1:])
 
-    steps_since_last_goal, avg_steps = check_if_goal(history, steps_since_last_goal, updated_entire_history)
+    model, steps_since_last_goal, avg_steps = check_if_goal(history, steps_since_last_goal, updated_entire_history, model)
 
     if (model is not None) and steps_since_last_goal > 3*avg_steps:
         log(f"Looks like we're stuck in a loop! steps_since_last_goal={steps_since_last_goal} and avg_steps={avg_steps}")
@@ -243,8 +72,8 @@ def process_history_sentinel(strData, entire_history, steps_since_last_goal):
 
     # If there is a trained model, use it to select the next action
     if (model is not None):
-        last_step = get_model_prediction(history)
-    
+        last_step = model.get_action(history)
+
     return updated_entire_history, last_step
 
 def main():
@@ -271,7 +100,8 @@ def main():
     entire_history = ''
     last_step = ''
     steps_since_last_goal = 0
-    
+    model = None
+
     try:
         with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as sock:
             sock.bind(('127.0.0.1', portNum))
@@ -281,7 +111,7 @@ def main():
                 log(f'Connected by {addr}')
                 while True:
                     data = conn.recv(1024)
-                    
+
                     while not data:
                         data = conn.recv(1024)
                         timeout += 10
@@ -289,7 +119,7 @@ def main():
                             log('ERROR: Timeout receiving next input from Java environment')
                             exit(-1)
                     # Connected to socket
-                    
+
                     # Handling socket communication
                     # Check for sentinel
                     strData = data.decode('utf-8')
@@ -301,7 +131,7 @@ def main():
                         conn.sendall('$$$ack'.encode('ASCII'))
 
                     elif (strData.startswith('$$$history:')):
-                        entire_history, last_step = process_history_sentinel(strData, entire_history, steps_since_last_goal)
+                        entire_history, last_step = process_history_sentinel(strData, entire_history, steps_since_last_goal, model)
                         # Send the model's prediction to the environment
                         last_step = send_letter(conn, last_step)
 
@@ -314,7 +144,7 @@ def main():
                         # Should never happen...
                         # if it does, send a random letter back to the Java environment
                         last_step = send_letter() 
-     
+
     except Exception as error:
         log('Exception:' + str(error))
         log('-----')
